_REFERENCE/.ralphex/integration/20260301183309/apps/api/requirements.txt
fastapi==0.116.1
sqlalchemy==2.0.38
alembic==1.16.5
psycopg[binary]==3.2.6
argon2-cffi==23.1.0
PyJWT==2.10.1
redis==6.4.0
rq==2.6.0
